Refactor #349 - migration core:component:uitext to core:uitext
