+++ app/build.gradle.kts
@@ -79,7 +79,6 @@ dependencies {
     implementation(projects.core.androidx.transition)
     implementation(projects.core.annotation)
     implementation(projects.core.component.tag)
-    implementation(projects.core.component.uitext)
     implementation(projects.core.dataDi)
     implementation(projects.core.data)
     implementation(projects.core.designsystem)
@@ -92,6 +91,7 @@ dependencies {
     implementation(projects.core.statekit)
     kapt(projects.statekit.compiler)
     implementation(projects.core.translation)
+    implementation(projects.core.uitext)
 
     implementation(libs.kotlinx.coroutines.android)
     implementation(libs.kotlinx.collections.immutable)
@@ -132,10 +132,10 @@ dependencies {
     testImplementation(projects.testkit)
     testImplementation(projects.statekit.test)
     testImplementation(projects.core.component.tagTest)
-    testImplementation(projects.core.component.uitextTest)
     testImplementation(projects.core.dataTest)
     testImplementation(projects.core.kotlinTest)
     testImplementation(projects.core.scheduleTest)
+    testImplementation(projects.core.uitextTest)
     androidTestImplementation(projects.core.local)
     androidTestImplementation(libs.kotlinx.coroutines.test)
     androidTestImplementation(libs.javafaker)
