+++ app/src/main/java/com/nlab/reminder/domain/feature/home/HomeUiState.kt
@@ -18,7 +18,7 @@ package com.nlab.reminder.domain.feature.home
 
 import com.nlab.reminder.core.data.model.Tag
 import com.nlab.reminder.core.kotlin.NonNegativeLong
-import com.nlab.reminder.core.component.text.UiText
+import com.nlab.reminder.core.text.UiText
 
 /**
  * @author Doohyun
