+++ app/src/test/java/com/nlab/reminder/domain/feature/home/HomeTestGenerator.kt
@@ -18,8 +18,8 @@ package com.nlab.reminder.domain.feature.home
 
 import com.nlab.reminder.core.component.tag.edit.TagEditDelegate
 import com.nlab.reminder.core.component.tag.edit.genTagEditState
-import com.nlab.reminder.core.component.text.UiText
-import com.nlab.reminder.core.component.text.genUiTexts
+import com.nlab.reminder.core.text.UiText
+import com.nlab.reminder.core.text.genUiTexts
 import com.nlab.reminder.core.data.model.Tag
 import com.nlab.reminder.core.data.model.genTags
 import com.nlab.reminder.core.data.repository.ScheduleRepository
