+++ core/uitext-test/src/main/AndroidManifest.xml
