+++ core/uitext/.gitignore
