+++ core/uitext/src/main/AndroidManifest.xml
