+++ core/uitext/src/main/kotlin/com/nlab/reminder/core/text/ui/UiTextExt.kt
@@ -14,16 +14,23 @@
  * limitations under the License.
  */
 
-package com.nlab.reminder.core.component.text
+package com.nlab.reminder.core.text.ui
 
-import com.nlab.testkit.faker.genBothify
-import com.nlab.testkit.faker.genInt
+import android.content.Context
+import com.nlab.reminder.core.text.UiText
 
 /**
  * @author Thalys
  */
-fun genUiText(value: String = genBothify()): UiText = UiText(value)
+fun UiText.toText(context: Context): String = when (this) {
+    is UiText.Direct -> value
+    is UiText.ResId -> {
+        if (args == null) context.getString(resId)
+        else context.getString(resId, *args)
+    }
 
-fun genUiTexts(size: Int = genInt(min = 1, max = 10)): List<UiText> = List(size) {
-    genUiText(value = "${genBothify()}_$it")
+    is UiText.PluralsResId -> {
+        if (args == null) context.resources.getQuantityString(resId, count)
+        else context.resources.getQuantityString(resId, count, *args)
+    }
 }
